Fix parsing and status of smartmonkey termination
SmartMonkey was reporting two states of termination
one as "Finsihed" and one as "Aborted" where both
are successful as compared to "Crashed". When tests
aborted due to a device state or UI exception, the
results were reported as not finished or crashed. This
required that we use intermediate or running time to
determine the true uptime. This had a parsing error
when reading the Starting time which was always zero.

Change-Id: Ia7e9396a2deda76e1dbc1ff33eecadbd0126fdb1