Make all parsers return null.
When a parser receives an empty input ("".equals(input.trim())), then
return null, not new IItem() with no input.  If there is input but
invalid, a non-null is still returned.

This fixes the case in bugreports where the section is listed but there
is no input.  For example:

    ----- SECTION 1 ------

    ----- SECTION 2 ------
    text

In this case, the section 1 parser would return null.

Change-Id: Ifdfabc3bc40aa2303fcbc11ec6ccbf63167d54ae